Update Provisioning and Editor
