+++ provisioning/dashboards/panels.json
@@ -109,7 +109,7 @@
         "overrides": []
       },
       "gridPos": {
-        "h": 19,
+        "h": 10,
         "w": 12,
         "x": 12,
         "y": 0
@@ -179,6 +179,7 @@
       "id": 4,
       "options": {
         "autoPlay": true,
+        "buttons": [],
         "controls": true,
         "height": 0,
         "heightMode": "auto",
@@ -213,6 +214,89 @@
       ],
       "title": "Image",
       "type": "volkovlabs-image-panel"
+    },
+    {
+      "datasource": {
+        "type": "marcusolsson-static-datasource",
+        "uid": "P1D2C73DC01F2359B"
+      },
+      "fieldConfig": {
+        "defaults": {
+          "color": {
+            "mode": "palette-classic"
+          },
+          "custom": {
+            "hideFrom": {
+              "legend": false,
+              "tooltip": false,
+              "viz": false
+            }
+          },
+          "mappings": []
+        },
+        "overrides": []
+      },
+      "gridPos": {
+        "h": 10,
+        "w": 12,
+        "x": 12,
+        "y": 10
+      },
+      "id": 7,
+      "options": {
+        "displayLabels": ["name"],
+        "legend": {
+          "displayMode": "table",
+          "placement": "right",
+          "showLegend": true,
+          "values": ["percent"]
+        },
+        "pieType": "pie",
+        "reduceOptions": {
+          "calcs": ["lastNotNull"],
+          "fields": "",
+          "values": true
+        },
+        "tooltip": {
+          "mode": "single",
+          "sort": "none"
+        }
+      },
+      "pluginVersion": "9.2.2",
+      "targets": [
+        {
+          "datasource": {
+            "type": "marcusolsson-static-datasource",
+            "uid": "P1D2C73DC01F2359B"
+          },
+          "frame": {
+            "fields": [
+              {
+                "config": {},
+                "name": "Name",
+                "type": "string",
+                "values": ["Graph", "Logs", "Node Graph", "Table", "Trace"]
+              },
+              {
+                "config": {},
+                "name": "Amount",
+                "type": "number",
+                "values": [10, 7, 3, 4, 2]
+              }
+            ],
+            "meta": {
+              "custom": {
+                "customCode": "const values = [['Graph', 'Logs', 'Node Graph', 'Table', 'Trace'], [10, 7, 3, 4, 2]];\n\nconst result = {\n  ...frame,\n  fields: frame.fields.map((field, i) => ({\n    ...field,\n    values: values[i]\n  }))\n}\n\nreturn Promise.resolve(result);",
+                "valuesEditor": "custom"
+              }
+            },
+            "name": "sales"
+          },
+          "refId": "A"
+        }
+      ],
+      "title": "Custom Code",
+      "type": "piechart"
     }
   ],
   "refresh": "",
