+++ src/constants/field.ts
@@ -1,6 +1,4 @@
 import { FieldType } from '@grafana/data';
-import { CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
-import { ValuesEditor } from '../types';
 
 /**
  * Field Types
@@ -19,35 +17,3 @@ export const FieldTypes = [
  * Length to show Text Area instead of Input
  */
 export const TextAreaLength = 100;
-
-/**
- * Values Editor Options
- */
-export const ValuesEditorOptions = [
-  {
-    label: 'Manual',
-    value: ValuesEditor.MANUAL,
-  },
-  {
-    label: 'Custom',
-    value: ValuesEditor.CUSTOM,
-  },
-];
-
-/**
- * Supported Languages
- */
-export const enum CodeLanguage {
-  JAVASCRIPT = 'javascript',
-}
-
-/**
- * Custom Values Editor Suggestions
- */
-export const CustomValuesEditorSuggestions: CodeEditorSuggestionItem[] = [
-  {
-    label: 'frame',
-    kind: CodeEditorSuggestionItemKind.Property,
-    detail: 'DataFrameDTO',
-  },
-];
