+++ src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from './datasource';
+export * from './editor';
 export * from './field';
 export * from './tests';
